discord_bot/requirements.txt
discord.py
slack_sdk
python-dotenv
psycopg2
